PLT-8148 Split validators into separate modules.
