+++ marlowe-plutus/test/Language/Marlowe/PlutusSpec.hs
@@ -57,11 +57,13 @@ import Language.Marlowe.Core.V1.Semantics.Types (
   getInputContent,
  )
 import qualified Language.Marlowe.Core.V1.Semantics.Types as M (Party (Address), State (..))
-import qualified Language.Marlowe.Plutus as Official (
-  marloweValidatorBytes,
-  marloweValidatorHash,
+import qualified Language.Marlowe.Plutus.RolePayout as Official (
   rolePayoutValidatorBytes,
   rolePayoutValidatorHash,
+ )
+import qualified Language.Marlowe.Plutus.Semantics as Official (
+  marloweValidatorBytes,
+  marloweValidatorHash,
  )
 import Language.Marlowe.Scripts.Types (MarloweInput, MarloweTxInput (..))
 import Paths_marlowe_plutus (getDataDir)
@@ -1278,179 +1280,179 @@ protocolVersion = MajorProtocolVersion 8
 -- | A default cost model for Plutus.
 costModel :: [(String, Integer)]
 costModel =
-  [ ("addInteger-cpu-arguments-intercept", 205665)
+  [ ("addInteger-cpu-arguments-intercept", 205_665)
   , ("addInteger-cpu-arguments-slope", 812)
   , ("addInteger-memory-arguments-intercept", 1)
   , ("addInteger-memory-arguments-slope", 1)
-  , ("appendByteString-cpu-arguments-intercept", 1000)
+  , ("appendByteString-cpu-arguments-intercept", 1_000)
   , ("appendByteString-cpu-arguments-slope", 571)
   , ("appendByteString-memory-arguments-intercept", 0)
   , ("appendByteString-memory-arguments-slope", 1)
-  , ("appendString-cpu-arguments-intercept", 1000)
-  , ("appendString-cpu-arguments-slope", 24177)
+  , ("appendString-cpu-arguments-intercept", 1_000)
+  , ("appendString-cpu-arguments-slope", 24_177)
   , ("appendString-memory-arguments-intercept", 4)
   , ("appendString-memory-arguments-slope", 1)
-  , ("bData-cpu-arguments", 1000)
+  , ("bData-cpu-arguments", 1_000)
   , ("bData-memory-arguments", 32)
-  , ("blake2b_256-cpu-arguments-intercept", 117366)
-  , ("blake2b_256-cpu-arguments-slope", 10475)
+  , ("blake2b_256-cpu-arguments-intercept", 117_366)
+  , ("blake2b_256-cpu-arguments-slope", 10_475)
   , ("blake2b_256-memory-arguments", 4)
-  , ("cekApplyCost-exBudgetCPU", 23000)
+  , ("cekApplyCost-exBudgetCPU", 23_000)
   , ("cekApplyCost-exBudgetMemory", 100)
-  , ("cekBuiltinCost-exBudgetCPU", 23000)
+  , ("cekBuiltinCost-exBudgetCPU", 23_000)
   , ("cekBuiltinCost-exBudgetMemory", 100)
-  , ("cekConstCost-exBudgetCPU", 23000)
+  , ("cekConstCost-exBudgetCPU", 23_000)
   , ("cekConstCost-exBudgetMemory", 100)
-  , ("cekDelayCost-exBudgetCPU", 23000)
+  , ("cekDelayCost-exBudgetCPU", 23_000)
   , ("cekDelayCost-exBudgetMemory", 100)
-  , ("cekForceCost-exBudgetCPU", 23000)
+  , ("cekForceCost-exBudgetCPU", 23_000)
   , ("cekForceCost-exBudgetMemory", 100)
-  , ("cekLamCost-exBudgetCPU", 23000)
+  , ("cekLamCost-exBudgetCPU", 23_000)
   , ("cekLamCost-exBudgetMemory", 100)
   , ("cekStartupCost-exBudgetCPU", 100)
   , ("cekStartupCost-exBudgetMemory", 100)
-  , ("cekVarCost-exBudgetCPU", 23000)
+  , ("cekVarCost-exBudgetCPU", 23_000)
   , ("cekVarCost-exBudgetMemory", 100)
-  , ("chooseData-cpu-arguments", 19537)
+  , ("chooseData-cpu-arguments", 19_537)
   , ("chooseData-memory-arguments", 32)
-  , ("chooseList-cpu-arguments", 175354)
+  , ("chooseList-cpu-arguments", 175_354)
   , ("chooseList-memory-arguments", 32)
-  , ("chooseUnit-cpu-arguments", 46417)
+  , ("chooseUnit-cpu-arguments", 46_417)
   , ("chooseUnit-memory-arguments", 4)
-  , ("consByteString-cpu-arguments-intercept", 221973)
+  , ("consByteString-cpu-arguments-intercept", 221_973)
   , ("consByteString-cpu-arguments-slope", 511)
   , ("consByteString-memory-arguments-intercept", 0)
   , ("consByteString-memory-arguments-slope", 1)
-  , ("constrData-cpu-arguments", 89141)
+  , ("constrData-cpu-arguments", 89_141)
   , ("constrData-memory-arguments", 32)
-  , ("decodeUtf8-cpu-arguments-intercept", 497525)
-  , ("decodeUtf8-cpu-arguments-slope", 14068)
+  , ("decodeUtf8-cpu-arguments-intercept", 497_525)
+  , ("decodeUtf8-cpu-arguments-slope", 14_068)
   , ("decodeUtf8-memory-arguments-intercept", 4)
   , ("decodeUtf8-memory-arguments-slope", 2)
-  , ("divideInteger-cpu-arguments-constant", 196500)
-  , ("divideInteger-cpu-arguments-model-arguments-intercept", 453240)
+  , ("divideInteger-cpu-arguments-constant", 196_500)
+  , ("divideInteger-cpu-arguments-model-arguments-intercept", 453_240)
   , ("divideInteger-cpu-arguments-model-arguments-slope", 220)
   , ("divideInteger-memory-arguments-intercept", 0)
   , ("divideInteger-memory-arguments-minimum", 1)
   , ("divideInteger-memory-arguments-slope", 1)
-  , ("encodeUtf8-cpu-arguments-intercept", 1000)
-  , ("encodeUtf8-cpu-arguments-slope", 28662)
+  , ("encodeUtf8-cpu-arguments-intercept", 1_000)
+  , ("encodeUtf8-cpu-arguments-slope", 28_662)
   , ("encodeUtf8-memory-arguments-intercept", 4)
   , ("encodeUtf8-memory-arguments-slope", 2)
-  , ("equalsByteString-cpu-arguments-constant", 245000)
-  , ("equalsByteString-cpu-arguments-intercept", 216773)
+  , ("equalsByteString-cpu-arguments-constant", 245_000)
+  , ("equalsByteString-cpu-arguments-intercept", 216_773)
   , ("equalsByteString-cpu-arguments-slope", 62)
   , ("equalsByteString-memory-arguments", 1)
-  , ("equalsData-cpu-arguments-intercept", 1060367)
-  , ("equalsData-cpu-arguments-slope", 12586)
+  , ("equalsData-cpu-arguments-intercept", 1_060_367)
+  , ("equalsData-cpu-arguments-slope", 12_586)
   , ("equalsData-memory-arguments", 1)
-  , ("equalsInteger-cpu-arguments-intercept", 208512)
+  , ("equalsInteger-cpu-arguments-intercept", 208_512)
   , ("equalsInteger-cpu-arguments-slope", 421)
   , ("equalsInteger-memory-arguments", 1)
-  , ("equalsString-cpu-arguments-constant", 187000)
-  , ("equalsString-cpu-arguments-intercept", 1000)
-  , ("equalsString-cpu-arguments-slope", 52998)
+  , ("equalsString-cpu-arguments-constant", 187_000)
+  , ("equalsString-cpu-arguments-intercept", 1_000)
+  , ("equalsString-cpu-arguments-slope", 52_998)
   , ("equalsString-memory-arguments", 1)
-  , ("fstPair-cpu-arguments", 80436)
+  , ("fstPair-cpu-arguments", 80_436)
   , ("fstPair-memory-arguments", 32)
-  , ("headList-cpu-arguments", 43249)
+  , ("headList-cpu-arguments", 43_249)
   , ("headList-memory-arguments", 32)
-  , ("iData-cpu-arguments", 1000)
+  , ("iData-cpu-arguments", 1_000)
   , ("iData-memory-arguments", 32)
-  , ("ifThenElse-cpu-arguments", 80556)
+  , ("ifThenElse-cpu-arguments", 80_556)
   , ("ifThenElse-memory-arguments", 1)
-  , ("indexByteString-cpu-arguments", 57667)
+  , ("indexByteString-cpu-arguments", 57_667)
   , ("indexByteString-memory-arguments", 4)
-  , ("lengthOfByteString-cpu-arguments", 1000)
+  , ("lengthOfByteString-cpu-arguments", 1_000)
   , ("lengthOfByteString-memory-arguments", 10)
-  , ("lessThanByteString-cpu-arguments-intercept", 197145)
+  , ("lessThanByteString-cpu-arguments-intercept", 197_145)
   , ("lessThanByteString-cpu-arguments-slope", 156)
   , ("lessThanByteString-memory-arguments", 1)
-  , ("lessThanEqualsByteString-cpu-arguments-intercept", 197145)
+  , ("lessThanEqualsByteString-cpu-arguments-intercept", 197_145)
   , ("lessThanEqualsByteString-cpu-arguments-slope", 156)
   , ("lessThanEqualsByteString-memory-arguments", 1)
-  , ("lessThanEqualsInteger-cpu-arguments-intercept", 204924)
+  , ("lessThanEqualsInteger-cpu-arguments-intercept", 204_924)
   , ("lessThanEqualsInteger-cpu-arguments-slope", 473)
   , ("lessThanEqualsInteger-memory-arguments", 1)
-  , ("lessThanInteger-cpu-arguments-intercept", 208896)
+  , ("lessThanInteger-cpu-arguments-intercept", 208_896)
   , ("lessThanInteger-cpu-arguments-slope", 511)
   , ("lessThanInteger-memory-arguments", 1)
-  , ("listData-cpu-arguments", 52467)
+  , ("listData-cpu-arguments", 52_467)
   , ("listData-memory-arguments", 32)
-  , ("mapData-cpu-arguments", 64832)
+  , ("mapData-cpu-arguments", 64_832)
   , ("mapData-memory-arguments", 32)
-  , ("mkCons-cpu-arguments", 65493)
+  , ("mkCons-cpu-arguments", 65_493)
   , ("mkCons-memory-arguments", 32)
-  , ("mkNilData-cpu-arguments", 22558)
+  , ("mkNilData-cpu-arguments", 22_558)
   , ("mkNilData-memory-arguments", 32)
-  , ("mkNilPairData-cpu-arguments", 16563)
+  , ("mkNilPairData-cpu-arguments", 16_563)
   , ("mkNilPairData-memory-arguments", 32)
-  , ("mkPairData-cpu-arguments", 76511)
+  , ("mkPairData-cpu-arguments", 76_511)
   , ("mkPairData-memory-arguments", 32)
-  , ("modInteger-cpu-arguments-constant", 196500)
-  , ("modInteger-cpu-arguments-model-arguments-intercept", 453240)
+  , ("modInteger-cpu-arguments-constant", 196_500)
+  , ("modInteger-cpu-arguments-model-arguments-intercept", 453_240)
   , ("modInteger-cpu-arguments-model-arguments-slope", 220)
   , ("modInteger-memory-arguments-intercept", 0)
   , ("modInteger-memory-arguments-minimum", 1)
   , ("modInteger-memory-arguments-slope", 1)
-  , ("multiplyInteger-cpu-arguments-intercept", 69522)
-  , ("multiplyInteger-cpu-arguments-slope", 11687)
+  , ("multiplyInteger-cpu-arguments-intercept", 69_522)
+  , ("multiplyInteger-cpu-arguments-slope", 11_687)
   , ("multiplyInteger-memory-arguments-intercept", 0)
   , ("multiplyInteger-memory-arguments-slope", 1)
-  , ("nullList-cpu-arguments", 60091)
+  , ("nullList-cpu-arguments", 60_091)
   , ("nullList-memory-arguments", 32)
-  , ("quotientInteger-cpu-arguments-constant", 196500)
-  , ("quotientInteger-cpu-arguments-model-arguments-intercept", 453240)
+  , ("quotientInteger-cpu-arguments-constant", 196_500)
+  , ("quotientInteger-cpu-arguments-model-arguments-intercept", 453_240)
   , ("quotientInteger-cpu-arguments-model-arguments-slope", 220)
   , ("quotientInteger-memory-arguments-intercept", 0)
   , ("quotientInteger-memory-arguments-minimum", 1)
   , ("quotientInteger-memory-arguments-slope", 1)
-  , ("remainderInteger-cpu-arguments-constant", 196500)
-  , ("remainderInteger-cpu-arguments-model-arguments-intercept", 453240)
+  , ("remainderInteger-cpu-arguments-constant", 196_500)
+  , ("remainderInteger-cpu-arguments-model-arguments-intercept", 453_240)
   , ("remainderInteger-cpu-arguments-model-arguments-slope", 220)
   , ("remainderInteger-memory-arguments-intercept", 0)
   , ("remainderInteger-memory-arguments-minimum", 1)
   , ("remainderInteger-memory-arguments-slope", 1)
-  , ("serialiseData-cpu-arguments-intercept", 1159724)
-  , ("serialiseData-cpu-arguments-slope", 392670)
+  , ("serialiseData-cpu-arguments-intercept", 1_159_724)
+  , ("serialiseData-cpu-arguments-slope", 392_670)
   , ("serialiseData-memory-arguments-intercept", 0)
   , ("serialiseData-memory-arguments-slope", 2)
-  , ("sha2_256-cpu-arguments-intercept", 806990)
-  , ("sha2_256-cpu-arguments-slope", 30482)
+  , ("sha2_256-cpu-arguments-intercept", 806_990)
+  , ("sha2_256-cpu-arguments-slope", 30_482)
   , ("sha2_256-memory-arguments", 4)
-  , ("sha3_256-cpu-arguments-intercept", 1927926)
-  , ("sha3_256-cpu-arguments-slope", 82523)
+  , ("sha3_256-cpu-arguments-intercept", 1_927_926)
+  , ("sha3_256-cpu-arguments-slope", 82_523)
   , ("sha3_256-memory-arguments", 4)
-  , ("sliceByteString-cpu-arguments-intercept", 265318)
+  , ("sliceByteString-cpu-arguments-intercept", 265_318)
   , ("sliceByteString-cpu-arguments-slope", 0)
   , ("sliceByteString-memory-arguments-intercept", 4)
   , ("sliceByteString-memory-arguments-slope", 0)
-  , ("sndPair-cpu-arguments", 85931)
+  , ("sndPair-cpu-arguments", 85_931)
   , ("sndPair-memory-arguments", 32)
-  , ("subtractInteger-cpu-arguments-intercept", 205665)
+  , ("subtractInteger-cpu-arguments-intercept", 205_665)
   , ("subtractInteger-cpu-arguments-slope", 812)
   , ("subtractInteger-memory-arguments-intercept", 1)
   , ("subtractInteger-memory-arguments-slope", 1)
-  , ("tailList-cpu-arguments", 41182)
+  , ("tailList-cpu-arguments", 41_182)
   , ("tailList-memory-arguments", 32)
-  , ("trace-cpu-arguments", 212342)
+  , ("trace-cpu-arguments", 212_342)
   , ("trace-memory-arguments", 32)
-  , ("unBData-cpu-arguments", 31220)
+  , ("unBData-cpu-arguments", 31_220)
   , ("unBData-memory-arguments", 32)
-  , ("unConstrData-cpu-arguments", 32696)
+  , ("unConstrData-cpu-arguments", 32_696)
   , ("unConstrData-memory-arguments", 32)
-  , ("unIData-cpu-arguments", 43357)
+  , ("unIData-cpu-arguments", 43_357)
   , ("unIData-memory-arguments", 32)
-  , ("unListData-cpu-arguments", 32247)
+  , ("unListData-cpu-arguments", 32_247)
   , ("unListData-memory-arguments", 32)
-  , ("unMapData-cpu-arguments", 38314)
+  , ("unMapData-cpu-arguments", 38_314)
   , ("unMapData-memory-arguments", 32)
-  , ("verifyEcdsaSecp256k1Signature-cpu-arguments", 35892428)
+  , ("verifyEcdsaSecp256k1Signature-cpu-arguments", 35_892_428)
   , ("verifyEcdsaSecp256k1Signature-memory-arguments", 10)
-  , ("verifyEd25519Signature-cpu-arguments-intercept", 57996947)
-  , ("verifyEd25519Signature-cpu-arguments-slope", 18975)
+  , ("verifyEd25519Signature-cpu-arguments-intercept", 57_996_947)
+  , ("verifyEd25519Signature-cpu-arguments-slope", 18_975)
   , ("verifyEd25519Signature-memory-arguments", 10)
-  , ("verifySchnorrSecp256k1Signature-cpu-arguments-intercept", 38887044)
-  , ("verifySchnorrSecp256k1Signature-cpu-arguments-slope", 32947)
+  , ("verifySchnorrSecp256k1Signature-cpu-arguments-intercept", 38_887_044)
+  , ("verifySchnorrSecp256k1Signature-cpu-arguments-slope", 32_947)
   , ("verifySchnorrSecp256k1Signature-memory-arguments", 10)
   ]
